+++ logservice/logpuller/region_request_scheduler.go
@@ -35,8 +35,8 @@ type regionRequestScheduler struct {
 	config   *SubscriptionClientConfig
 	upstream *upstreamHandle
 
-	pushRegionEventToDS func(SubscriptionID, regionEvent)
-	onRegionFail        func(regionErrorInfo)
+	eventSink       *regionEventSink
+	failureReporter *regionFailureReporter
 
 	queue *priorityqueue.PriorityQueue[*regionPriorityTask]
 	seq   atomic.Uint64
@@ -46,11 +46,11 @@ type regionRequestScheduler struct {
 
 func newRegionRequestScheduler(client *subscriptionClient) *regionRequestScheduler {
 	return &regionRequestScheduler{
-		config:              client.config,
-		upstream:            client.upstream,
-		pushRegionEventToDS: client.pushRegionEventToDS,
-		onRegionFail:        client.onRegionFail,
-		queue:               priorityqueue.New[*regionPriorityTask](),
+		config:          client.config,
+		upstream:        client.upstream,
+		eventSink:       client.eventSink,
+		failureReporter: client.failureReporter,
+		queue:           priorityqueue.New[*regionPriorityTask](),
 	}
 }
 
@@ -103,7 +103,7 @@ func (s *regionRequestScheduler) attachRPCContextForRegion(ctx context.Context, 
 			zap.Uint64("regionID", region.verID.GetID()),
 			zap.Error(err))
 	}
-	s.onRegionFail(newRegionErrorInfo(region, &rpcCtxUnavailableErr{verID: region.verID}))
+	s.failureReporter.Report(newRegionErrorInfo(region, &rpcCtxUnavailableErr{verID: region.verID}))
 	return region, false
 }
 
@@ -150,8 +150,8 @@ func (s *regionRequestScheduler) Run(ctx context.Context, eg *errgroup.Group) er
 				rs,
 				perWorkerQueueSize,
 				s.upstream,
-				s.pushRegionEventToDS,
-				s.onRegionFail,
+				s.eventSink,
+				s.failureReporter,
 			)
 			rs.requestWorkers = append(rs.requestWorkers, requestWorker)
 		}
